MCOL-4466 regr_count() and distinct_count() now use long long for counters and for a type cast to/from boost::any

+++ utils/regr/regr_count.cpp
@@ -38,7 +38,7 @@ static Add_regr_count_ToUDAFMap addToMap;
 // Use the simple data model
 struct regr_count_data
 {
-    uint64_t	cnt;
+    long long	cnt;
 };
